+++ src/algebra/mod.rs
@@ -1,4 +1,4 @@
 pub mod std_io;
-pub mod dirs;
+pub mod user_dirs;
 pub mod symlinks;
 pub mod hop;
